Move serving requests out of getdns_query
