+++ src/test/Makefile.in
@@ -66,7 +66,9 @@ CHECK_CFLAGS=@CHECK_CFLAGS@
 CHECK_OBJS=check_getdns_common.lo check_getdns_context_set_timeout.lo \
 	check_getdns.lo check_getdns_transport.lo
 
-DECOMPOSED_OBJS_WITHOUT_JSMN=getdns_str2dict.lo getdns_context_config.lo
+DECOMPOSED_OBJS_WITHOUT_JSMN=getdns_str2dict.lo getdns_context_config.lo \
+	getdns_context_set_listen_addresses.lo
+
 DECOMPOSED_OBJS=$(DECOMPOSED_OBJS_WITHOUT_JSMN) jsmn.lo
 
 ALL_OBJS=$(CHECK_OBJS) check_getdns_libevent.lo check_getdns_libev.lo \
@@ -273,9 +275,14 @@ check_getdns_transport.lo check_getdns_transport.o: $(srcdir)/check_getdns_trans
  ../getdns/getdns_extra.h
 getdns_context_config.lo getdns_context_config.o: $(srcdir)/getdns_context_config.c $(srcdir)/getdns_context_config.h \
  ../getdns/getdns.h ../getdns/getdns_extra.h
+getdns_context_set_listen_addresses.lo getdns_context_set_listen_addresses.o: \
+ $(srcdir)/getdns_context_set_listen_addresses.c ../config.h \
+ $(srcdir)/getdns_context_set_listen_addresses.h ../getdns/getdns.h \
+ ../getdns/getdns_extra.h $(srcdir)/../types-internal.h ../getdns/getdns.h \
+ ../getdns/getdns_extra.h $(srcdir)/../util/rbtree.h
 getdns_query.lo getdns_query.o: $(srcdir)/getdns_query.c ../config.h $(srcdir)/../debug.h ../config.h \
  $(srcdir)/getdns_str2dict.h ../getdns/getdns.h $(srcdir)/getdns_context_config.h \
- ../getdns/getdns_extra.h
+ $(srcdir)/getdns_context_set_listen_addresses.h ../getdns/getdns_extra.h
 getdns_str2dict.lo getdns_str2dict.o: $(srcdir)/getdns_str2dict.c ../config.h $(srcdir)/../const-info.h \
  $(srcdir)/jsmn/jsmn.h $(srcdir)/getdns_str2dict.h ../getdns/getdns.h $(srcdir)/../types-internal.h \
  ../getdns/getdns.h ../getdns/getdns_extra.h $(srcdir)/../util/rbtree.h $(srcdir)/../list.h \
